Improves data normalization and training volume
Ensures correct normalization by casting the anchor peak length to a float.

Increases the minimum training volume for FDR estimation to improve statistical robustness, using the larger of the old value and a new minimum value.

+++ AlgorithmsFFLib/src/CandidateScorertron.cpp
@@ -1873,7 +1873,7 @@ namespace {
         e = ErrorUtils::isFalse(anchorPeakLength <= 0); ree;
 
         for (int i = 0; i < peakLengths.size(); i++) {
-            candidateScores->featuresArray[Features::MzPeakLengthsNorm1 + i] = peakLengths[i] / anchorPeakLength;
+            candidateScores->featuresArray[Features::MzPeakLengthsNorm1 + i] = peakLengths[i] / static_cast<float>(anchorPeakLength);
         }
 
         candidateScores->featuresArray[MzPeakLengthsMean] = MathUtils::mean(peakLengths);
